feat(nav): make map the home screen, move feed to /feed
